# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Stash-box metadata database",
        "title": "Stash-box REST API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Stash App",
            "url": "https://github.com/stashapp/stash-box"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/scenes/edits/{editId}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "scenes"
                ],
                "summary": "Update a pending scene edit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Edit ID (UUID)",
                        "name": "editId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated scene edit data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_rest_handlers_scene.UpdateSceneEditRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.Edit"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/scenes/{id}/edits": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "scenes"
                ],
                "summary": "Create a new scene edit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scene ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Scene edit data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_rest_handlers_scene.CreateSceneEditRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.Edit"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_gofrs_uuid.NullUUID": {
            "type": "object",
            "properties": {
                "uuid": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
...
        "github_com_stashapp_stash-box_internal_models.Edit": {
            "type": "object",
            "properties": {
                "applied": {
                    "type": "boolean"
                },
                "bot": {
                    "type": "boolean"
                },
                "closed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "string"
                },
                "operation": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "target_type": {
                    "type": "string"
                },
                "update_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "$ref": "#/definitions/github_com_gofrs_uuid.NullUUID"
                },
                "votes": {
                    "type": "integer"
                }
            }
        },
...
        "internal_api_rest_handlers_scene.CreateSceneEditRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "director": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "fingerprints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api_rest_handlers_scene.FingerprintInput"
                    }
                },
                "performer_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "studio_id": {
                    "type": "string"
                },
                "tag_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "internal_api_rest_handlers_scene.FingerprintInput": {
            "type": "object",
            "properties": {
                "algorithm": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "hash": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_rest_handlers_scene.UpdateSceneEditRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "director": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "fingerprints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api_rest_handlers_scene.FingerprintInput"
                    }
                },
                "performer_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "studio_id": {
                    "type": "string"
                },
                "tag_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key authentication",
            "type": "apiKey",
            "name": "ApiKey",
            "in": "header"
        },
        "SessionAuth": {
            "description": "Session cookie authentication",
            "type": "apiKey",
            "name": "stashbox",
            "in": "cookie"
        }
    }
}